response/requirements.txt
fastapi
uvicorn
PyYAML
httpx
python-iptables
pydantic
